move vmStatus logic to utils/status/vm (#264)
* move vmStatus logic to utils/status/vm

* move selection logic from vm status to selectors
