+++ src/components/VmStatus/index.js
@@ -1,2 +1,1 @@
 export { VmStatus, VmStatuses } from './VmStatus';
-export * from './getVmStatus';
