+++ src/selectors/common.js
@@ -6,3 +6,7 @@ export const getUid = resource => get(resource, 'metadata.uid');
 export const getId = value => `${getNamespace(value)}-${getName(value)}`;
 
 export const getStorageSize = resources => get(resources, 'requests.storage');
+
+export const getStatusConditions = entity => get(entity, 'status.conditions', []);
+export const getStatusConditionOfType = (entity, type) =>
+  getStatusConditions(entity).find(condition => condition.type === type);
