+++ src/selectors/index.js
@@ -1,6 +1,7 @@
 export * from './common';
 export * from './dv';
 export * from './host';
+export * from './migration';
 export * from './pod';
 export * from './pvc';
 export * from './vm';
