+++ src/selectors/migration/index.js
@@ -1,2 +1,1 @@
 export * from './selectors';
-export * from './container';
