+++ src/selectors/vm/combined.js
@@ -1,0 +1,4 @@
+import { isVmiRunning } from '../vmi';
+import { isVmRunning } from './selectors';
+
+export const isVmStarting = (vm, vmi) => isVmRunning(vm) && !isVmiRunning(vmi);
