+++ src/selectors/vm/index.js
@@ -1,1 +1,2 @@
+export * from './combined';
 export * from './selectors';
