+++ src/selectors/vmi/selectors.js
@@ -1,5 +1,7 @@
 import { get } from 'lodash';
 
+export const isVmiRunning = vmi => get(vmi, 'status.phase') === 'Running';
+
 export const getVmiIpAddresses = vmi =>
   get(vmi, 'status.interfaces', [])
     // get IPs only for named interfaces because Windows reports IPs for other devices like Loopback Pseudo-Interface 1 etc.
