+++ src/utils/status/vm/vmStatus.js
@@ -1,5 +1,3 @@
-import { get, includes } from 'lodash';
-
 import {
   VM_STATUS_POD_ERROR,
   VM_STATUS_ERROR,
@@ -12,21 +10,25 @@ import {
   VM_STATUS_VMI_WAITING,
   VM_STATUS_UNKNOWN,
   VM_STATUS_OTHER,
-} from '../../constants';
+} from './constants';
+
+import {
+  isPodSchedulable,
+  getStatusConditions,
+  getContainerStatuses,
+  isContainerFailing,
+  isMigrating,
+  isVmRunning,
+  isVmReady,
+  isVmCreated,
+  getContainerStatusReason,
+  getMigrationStatusPhase,
+} from '../../../selectors';
 
 const NOT_HANDLED = null;
 
-const failingContainerStatus = ['ImagePullBackOff', 'ErrImagePull', 'CrashLoopBackOff'];
-
-const isMigrationStatus = (migration, status) => {
-  const phase = get(migration, 'status.phase');
-  return phase && phase.toLowerCase() === status.toLowerCase();
-};
-
-const getConditionOfType = (pod, type) => get(pod, 'status.conditions', []).find(condition => condition.type === type);
-
 const getNotRedyConditionMessage = pod => {
-  const notReadyCondition = get(pod, 'status.conditions', []).find(condition => condition.status !== 'True');
+  const notReadyCondition = getStatusConditions(pod).find(condition => condition.status !== 'True');
   if (notReadyCondition) {
     // at least one pod condition not met. This can be just tentative, let the user analyze progress via Pod events
     return notReadyCondition.message || `Step: ${notReadyCondition.type}`;
@@ -34,48 +36,48 @@ const getNotRedyConditionMessage = pod => {
   return undefined;
 };
 
-const getFailingContainerStatus = pod =>
-  get(pod, 'status.containerStatuses', []).find(
-    container => !container.ready && includes(failingContainerStatus, get(container, 'state.waiting.reason'))
-  );
+const findFailingContainerStatus = pod => getContainerStatuses(pod).find(isContainerFailing);
 
-const getContainerStatusReason = containerStatus => {
-  const status = Object.getOwnPropertyNames(containerStatus.state).find(pn => !!containerStatus.state[pn].reason);
-  return status ? containerStatus.state[status].message : undefined;
-};
-
-const isSchedulable = pod => {
-  const podScheduledCond = getConditionOfType(pod, 'PodScheduled');
-  return !(podScheduledCond && podScheduledCond.status !== 'True' && podScheduledCond.reason === 'Unschedulable');
-};
-
-export const isRunning = vm => {
-  if (!get(vm, 'spec.running', false)) {
-    return { status: VM_STATUS_OFF };
+const isBeingMigrated = (vm, migration) => {
+  if (isMigrating(migration)) {
+    return { status: VM_STATUS_MIGRATING, message: getMigrationStatusPhase(migration) };
   }
-  // spec.running === true
   return NOT_HANDLED;
 };
 
+const isRunning = vm => (isVmRunning(vm) ? NOT_HANDLED : { status: VM_STATUS_OFF });
+
 const isReady = vm => {
-  if (get(vm, 'status.ready', false)) {
+  if (isVmReady(vm)) {
     // we are all set
     return { status: VM_STATUS_RUNNING };
   }
   return NOT_HANDLED;
 };
 
+const isVmError = vm => {
+  // is an issue with the VM definition?
+  const condition = getStatusConditions(vm)[0];
+  if (condition) {
+    // Do we need to analyze additional conditions in the array? Probably not.
+    if (condition.type === 'Failure') {
+      return { status: VM_STATUS_ERROR, message: condition.message };
+    }
+  }
+  return NOT_HANDLED;
+};
+
 const isCreated = (vm, launcherPod = null) => {
-  if (get(vm, 'status.created', false)) {
+  if (isVmCreated(vm)) {
     // created but not yet ready
     let message;
     if (launcherPod) {
       // pod created, so check for it's potential error
-      if (!isSchedulable(launcherPod)) {
+      if (!isPodSchedulable(launcherPod)) {
         return { status: VM_STATUS_POD_ERROR, message: 'Pod scheduling failed.' };
       }
 
-      const failingContainer = getFailingContainerStatus(launcherPod);
+      const failingContainer = findFailingContainerStatus(launcherPod);
       if (failingContainer) {
         return { status: VM_STATUS_POD_ERROR, message: getContainerStatusReason(failingContainer) };
       }
@@ -87,22 +89,10 @@ const isCreated = (vm, launcherPod = null) => {
   return NOT_HANDLED;
 };
 
-const isVmError = vm => {
-  // is an issue with the VM definition?
-  const condition = get(vm, 'status.conditions[0]');
-  if (condition) {
-    // Do we need to analyze additional conditions in the array? Probably not.
-    if (condition.type === 'Failure') {
-      return { status: VM_STATUS_ERROR, message: condition.message };
-    }
-  }
-  return NOT_HANDLED;
-};
-
 const isBeingImported = (vm, importerPods) => {
-  if (importerPods && importerPods.length > 0 && !get(vm, 'status.created', false)) {
+  if (importerPods && importerPods.length > 0 && !isVmCreated(vm)) {
     const importerPodsStatuses = importerPods.map(pod => {
-      if (!isSchedulable(pod)) {
+      if (!isPodSchedulable(pod)) {
         return {
           status: VM_STATUS_IMPORT_ERROR,
           message: 'Importer pod scheduling failed.',
@@ -110,7 +100,7 @@ const isBeingImported = (vm, importerPods) => {
         };
       }
 
-      const failingContainer = getFailingContainerStatus(pod);
+      const failingContainer = findFailingContainerStatus(pod);
       if (failingContainer) {
         return {
           status: VM_STATUS_IMPORT_ERROR,
@@ -139,24 +129,15 @@ const isBeingImported = (vm, importerPods) => {
   return NOT_HANDLED;
 };
 
-export const isBeingMigrated = (vm, migration) => {
-  if (migration) {
-    if (!isMigrationStatus(migration, 'succeeded') && !isMigrationStatus(migration, 'failed')) {
-      return { status: VM_STATUS_MIGRATING, message: get(migration, 'status.phase') };
-    }
-  }
-  return NOT_HANDLED;
-};
-
 const isWaitingForVmi = vm => {
   // assumption: spec.running === true
-  if (!get(vm, 'status.created', false)) {
+  if (!isVmCreated(vm)) {
     return { status: VM_STATUS_VMI_WAITING };
   }
   return NOT_HANDLED;
 };
 
-export const getVmStatusDetail = (vm, launcherPod, importerPods, migration) =>
+export const getVmStatus = (vm, launcherPod, importerPods, migration) =>
   isBeingMigrated(vm, migration) || // must be precceding isRunning() since vm.status.ready is true for a migrating VM
   isRunning(vm) ||
   isReady(vm) ||
@@ -165,10 +146,7 @@ export const getVmStatusDetail = (vm, launcherPod, importerPods, migration) =>
   isBeingImported(vm, importerPods) ||
   isWaitingForVmi(vm) || { status: VM_STATUS_UNKNOWN };
 
-export const getVmStatus = (vm, launcherPod, importerPods, migration) => {
-  const vmStatus = getVmStatusDetail(vm, launcherPod, importerPods, migration).status;
+export const getSimpleVmStatus = (vm, launcherPod, importerPods, migration) => {
+  const vmStatus = getVmStatus(vm, launcherPod, importerPods, migration).status;
   return vmStatus === VM_STATUS_OFF || vmStatus === VM_STATUS_RUNNING ? vmStatus : VM_STATUS_OTHER;
 };
-
-export const isVmiRunning = vmi => get(vmi, 'status.phase') === 'Running';
-export const isVmStarting = (vm, vmi) => get(vm, 'spec.running') && !isVmiRunning(vmi);
